wrappercheck: augment output of daemon with time stamps
Useful for correlating events in the daemon with events in testing.
We need to use a process group to deliver SIGINT/SIGTERM, otherwise
we cannot be sure that we catch all processes created by the daemon.

The return code of the daemon was not checked before (accidentally?!)
and this patch does not change that. Might be fixed in the future.

+++ test/wrappercheck.sh
@@ -59,7 +59,11 @@ if [ "$DAEMON_LOG" ] && [ "$WAIT_FOR_DAEMON_OUTPUT" ]; then
 fi
 
 ( set +x; echo >&2 "*** starting ${BACKGROUND[0]} as background daemon, output to ${DAEMON_LOG:-stderr}" )
-( set -x; exec >>${DAEMON_LOG:-&2} 2>&1; exec env "${ENV[@]}" "${BACKGROUND[@]}" ) &
+# We need to create a process group so that we can kill all processes started by the sub-shell.
+# ${BACKGROUND[*]} is used instead of ${BACKGROUND[@]} because although the later should have
+# avoided expansion of words (good!) somehow the quoting got messed up in practice (bad!).
+( set -x; exec >>${DAEMON_LOG:-&2} 2>&1; exec env "${ENV[@]}" setsid /bin/bash -c "set -x -o pipefail; ${BACKGROUND[*]} | $(dirname $0)/logger.py" ) &
+
 BACKGROUND_PID=$!
 PIDS+="$BACKGROUND_PID"
 
@@ -89,18 +93,27 @@ else
 fi
 
 ( set +x; echo >&2 "*** killing and waiting for ${BACKGROUND[0]}" )
-kill -INT $BACKGROUND_PID && kill -TERM $BACKGROUND_PID || true
-perl -e "sleep(60); kill(9, $BACKGROUND_PID);" &
-KILL_PID=$!
+if kill -INT -$BACKGROUND_PID 2>/dev/null && kill -TERM -$BACKGROUND_PID 2>/dev/null; then
+    perl -e "sleep(60); kill(9, -$BACKGROUND_PID);" &
+    KILL_PID=$!
+else
+    KILL_PID=
+fi
 set +e
 wait $BACKGROUND_PID
-msg=$(LC_ALL=C kill -KILL $KILL_PID 2>&1)
 SUBRET=$?
-if echo "$msg" | grep -q 'No such process'; then
-    # Consider this a success.
-    SUBRET=0
-else
-    echo "$msg"
+case $SUBRET in 0|130|137|139|143) SUBRET=0;; # 130 and 143 indicate that it was killed, probably by us, which is okay
+esac
+SUBRET=0 # TODO: don't ignore daemon results
+if [ "$KILL_PID" ]; then
+    msg=$(LC_ALL=C kill -KILL $KILL_PID 2>&1)
+    if echo "$msg" | grep -q 'No such process'; then
+        # Consider this a success.
+        SUBRET=0
+    else
+        echo "$msg"
+    fi
+    wait $KILL_PID
 fi
 set -e
 if [ $RET = 0 ]; then
